refactor: reorganize product-related models and update imports for better structure
